pyproject.toml
[project]
name = "mas"
version = "0.2.0"
description = "CLI-governed delivery scaffold for Claude Code sessions"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "click>=8.1",
]

[project.scripts]
mas = "core.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["mas/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mas/core"]
